third_party/blink/web_tests/virtual/clear-site-data-client-hints-support/external/wpt/client-hints/clear-site-data/clear-site-data-cache.https-expected.txt

This is a testharness.js-based test.
FAIL Clear-Site-Data for cache should remove all client hints. assert_equals: expected "MissingDeviceMemory" but got "HadDeviceMemory"
Harness: the test ran to completion.
